+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add Fingerbot support (`szjqr`).
 - Add Smart Lock support (`ms`, `jtmspro`). (#120) Thanks @pfgimutao for the contribution
 - Add Alarm Host support (`mal`). (#246) Thanks @bFollon for the contribution
+- Add Vibration Sensor support (`zd`).
 
 
 ### Fixed